- Spawn Handler: Add the remaining items.
I think this is everything? Still gotta make a menu but it's functional at least. This is incredibly tiresome so gonna be happy to do the menu.

Spawn chances also are just completely arbitrary and usually 14, I don't have the mental capacity nor time to test all of these individually yet to get perfect spawn chances for it and admittedly I hope most users are going to configure the chances individually.